+++ src/pie_manager/fileio.py
@@ -64,10 +64,11 @@ def get_config():
 
 changelog = f"""
 {Fore.GREEN}Pie Project Manager{Fore.RESET}
-{Fore.GREEN}v1.0.0{Style.RESET_ALL}
+{Fore.GREEN}v1.0.1{Style.RESET_ALL}
 
-{Fore.GREEN}+ Official release on pypi{Fore.RESET}
-{Fore.RED}- Will no longer be hosted on test pypi{Fore.RESET}
+{Fore.GREEN}+ Updated code for package check to make it more accurate <sysreqs.py>{Fore.RESET}
+{Fore.GREEN}+ Added yaspin and colorama to package check <sysreqs.py>{Fore.RESET}
+{Fore.GREEN}+ Fixed issue where deleting project files which contains readonly files inside will crash the program(Windows) <cli.py>{Fore.RESET}
 ----------------
 {Fore.GREEN}For full changelogs, visit https://github.com/skandabhairava/Pie-manager/blob/main/CHANGELOG.md{Fore.RESET}
 """