+++ src/pie_manager/sysreqs.py
@@ -1,21 +1,37 @@
 import subprocess, os, sys
 
 def test_package():
-    packages = ["pip", "pipreqs"]
+    packages = []
 
     try:
         import pip
+    except Exception as e:
+        packages.append("pip")
+
+    try:
         import pipreqs
     except Exception as e:
+        packages.append("pipreqs")
+
+    try:
+        import yaspin
+    except Exception as e:
+        packages.append("yaspin")
+
+    try:
+        import colorama
+    except Exception as e:
+        packages.append("colorama")
+
+    if len(packages) > 0:
         print("Some packages are not installed. Please install them before continuing.")
         print("Please install these packages:\n", "\n ".join(packages))
         sys.exit(1)
 
+    packages_not_installed = [x for x in ([["git", "--version"]]) if subprocess.run([*x]).returncode != 0]
     if os.name == 'nt':
-        packages_not_installed = [x for x in ([["git", "--version"]]) if subprocess.run([*x]).returncode != 0]
         os.system("cls")
     else:
-        packages_not_installed = [x for x in ([["git", "--version"]]) if subprocess.run([*x]).returncode != 0]
         os.system("clear")
 
     if packages_not_installed: